+++ R-package/R/sumstats.R
@@ -197,8 +197,8 @@ print.Sumstats <- function(x, ...) {
   pos <- 0L
   offsets <- data.frame(
     shard_label = character(length(shard_list)),
-    start0 = numeric(length(shard_list)),
-    stop0 = numeric(length(shard_list)),
+    start0 = integer(length(shard_list)),
+    stop0 = integer(length(shard_list)),
     stringsAsFactors = FALSE
   )
   for (i in seq_along(shard_list)) {
